pyproject.toml
[tool.poetry]
name = "poedss"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.4"
python-dateutil = "^2.8.2"
requests3 = "^0.0.0"
requests = "^2.27.1"
pandas = "^1.4.2"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"

[tool.poetry.scripts]
poedss = "poedss.poedss:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
